// webSpa/src/core/enums/enums.ts
export enum EnumChoice {
    No = 0,
    Yes = 1
}

export enum EnumTypeMessage {
    Personalized = 0,
    New = 1,
    Edit = 2,
    Delete = 3,
    Export = 4,
    Token = 5,
    TokenError = 6,
    TokenExpired = 7,
    FormValid = 8,
    Login = 9,
    Sync = 10,
    SyncError = 11
}

export enum EnumActionMessage {
    Success = 1,
    Error = 2,
    Info = 3,
    Warning = 4
}

export enum EnumHint {
    Numbers = 0,
    Date = 1,
    Hour = 2,
    Calc = 3,
    Separator = 4,
    ScoreBoard = 5,
    ViewField = 6
}

export enum EnumTypeAction {
    Menu = 0,
    Others = 1,
    Create = 2,
    Research = 3,
    Update = 4,
    Delete = 5,
    Authenticate = 6,
    View = 7,
    Disable = 8
}

export enum EnumUserStatus {
    Inactive = 0,
    Active = 1,
}